refactor: move posts ui cards into features with stubs
